:recycle: [Refactor]: Reservation List 컴포넌트 분리
