+++ 42manito/src/pages/index.tsx
@@ -7,7 +7,7 @@ import { initAllMentor } from "@/RTK/Slices/Home";
 import { signIn } from "@/RTK/Slices/Global";
 import { useMentorModal } from "@/hooks/Mentor/MentorModal";
 import { useFetchHome } from "@/hooks/Home/FetchHome";
-import ReservationRequests from "@/components/Home/Reservation/ReservationRequests";
+import ReservationLists from "@/components/Reservation/ReservationLists";
 import TopBanner from "@/components/Global/TopBanner";
 import { useGetCategoriesQuery } from "@/RTK/Apis/Category";
 import CategoryIconList from "@/components/Home/CategoryIconList";
@@ -16,13 +16,15 @@ import { Spin } from "antd";
 import MentorCard from "@/components/Mentor/Card";
 import { MentorProfileDto } from "@/Types/MentorProfiles/MentorProfile.dto";
 import ReservationModal from "@/components/Reservation/modal/ReservationModal";
+import { useGetRequestReservationsQuery } from "@/RTK/Apis/Reservation";
+import { ReservationDefaultDto } from "@/Types/Reservations/ReservationDefault.dto";
 
 const MentorModal = dynamic(() => import("@/components/Mentor/Modal"));
 
 export default function Home() {
   const dispatch = useAppDispatch();
   const OwnerId = useSelector(
-    (state: RootState) => state.rootReducers.global.uId
+    (state: RootState) => state.rootReducers.global.uId,
   );
   const [hasMore, setHasMore] = React.useState<boolean>(true);
   const [categoryId, setCategoryId] = React.useState<number>(0);
@@ -31,7 +33,7 @@ export default function Home() {
   const { newMentor, fetchNewCategory, fetchMoreData } =
     useFetchHome(categoryId);
   const [mentorList, setMentorList] = useState<MentorProfileDto[] | undefined>(
-    undefined
+    undefined,
   );
   const isModalOpen = useSelector(
     (state: RootState) => state.rootReducers.reservation.isModalOpen,
@@ -39,6 +41,22 @@ export default function Home() {
   const selectedReservation = useSelector(
     (state: RootState) => state.rootReducers.reservation.selectedReservation,
   );
+  const userId = useSelector(
+    (state: RootState) => state.rootReducers.global.uId,
+  );
+  const { data: response } = useGetRequestReservationsQuery({ id: userId });
+  const [reservations, setReservationRequests] = useState<
+    ReservationDefaultDto[]
+  >([]);
+
+  useEffect(() => {
+    if (response) {
+      setReservationRequests([
+        ...response.menteeReservations,
+        ...response.mentorReservations,
+      ]);
+    }
+  }, [response]);
 
   useEffect(() => {
     if (OwnerId === 0) {
@@ -109,7 +127,10 @@ export default function Home() {
                 <div className="home-text-detail">
                   멘토링 요청들을 확인해보세요
                 </div>
-                <ReservationRequests />
+                <ReservationLists
+                  reservations={reservations}
+                  emptyMsg={"대기 중인 멘토링이 없습니다 🥲"}
+                />
               </div>
             )}
           </div>
